raspberry/requirements.txt
pyserial
flask
flask-socketio
eventlet
python-dotenv

